+++ src/main/java/com/echobox/api/linkedin/types/LinkedInURNIdType.java
@@ -19,7 +19,6 @@ package com.echobox.api.linkedin.types;
 
 import com.echobox.api.linkedin.jsonmapper.LinkedIn;
 import com.echobox.api.linkedin.types.urn.URN;
-
 import lombok.Getter;
 
 /**