+++ src/main/java/com/echobox/api/linkedin/types/v1/Company.java
@@ -19,7 +19,6 @@ package com.echobox.api.linkedin.types.v1;
 
 import com.echobox.api.linkedin.jsonmapper.LinkedIn;
 import com.echobox.api.linkedin.types.LinkedInIdAndNameType;
-
 import lombok.Getter;
 import lombok.Setter;