+++ src/main/java/com/echobox/api/linkedin/types/v1/IndustryCode.java
@@ -18,9 +18,7 @@
 package com.echobox.api.linkedin.types.v1;
 
 import com.echobox.api.linkedin.logging.LinkedInLogger;
-
 import lombok.Getter;
-
 import org.slf4j.Logger;
 
 /**