+++ src/test/java/com/echobox/api/linkedin/jsonmapper/SimpleJSON.java
@@ -18,7 +18,6 @@
 package com.echobox.api.linkedin.jsonmapper;
 
 import com.echobox.api.linkedin.types.v1.CodeAndNameType;
-
 import lombok.Getter;
 
 import java.util.List;