+++ CHANGELOG.md
@@ -36,7 +36,7 @@
   representation of the urn. Instead of using `URNEntityType.name()`
   `URNEntityType.getEntityValue()` should be used to support UGCPosts.
 
-## 1.0.3 (Work in progress)
+## 1.0.3 (May 18, 2020)
 *  `DefaultLinkedInClient.makeRequestAndProcessResponse` handles 401
    errors as `LinkedInOAuthException` rather than
    `LinkedInNetworkException`
@@ -54,4 +54,26 @@
   videos. LinkedIn is planning to deprecate the existing Rich Media 
   Platform by `Janurary 30, 2020`.
   See [migration guide](https://docs.microsoft.com/en-us/linkedin/shared/references/migrations/rich-media-platform-deprecation?context=linkedin/marketing/context&trk=eml-mktg-20191028-developer-email-api-updates-october&mcid=6592215409070530560&src=e-eml).
-   
+* Add `NATIVE_DOCUMENT`, `URN_REFERENCE`, `LIVE_VIDEO` to UGC ShareMediaCategory enum.
+* Checkstyle plugin version bumped. Some code reformatted in order to satisfy new checkstyle rules.
+
+## 1.0.4 (June 15, 2020)
+* Put in a place an interim fix for Linkedin V2 pagination. It seems like LinkedIn's pagination 
+is incorrectly returning the number of elements where the number of elements is less than the 
+request amount even though there are in fact more results. This fix should continue to 
+the next page until there are no more results.
+
+## 1.0.5 (June 18, 2020)
+* Accidentally skipped this version - no changes
+
+## 1.0.6 (June 18, 2020)
+* After receiving a response from LinkedIn on the issues surrounding V2 pagination where the 
+`count` value in the response object does not match the number of elements returned or in fact in
+some cases where not all elements are returned, it's best to not provide pagination parameters in
+the request to the LinkedIn API until they have fixed it in the API. Pagination parameters will not
+longer be provided by default in `DefaultLinkedInClient.fetchConnection()`. Instead, if no `count`
+parameter is provided in the initial request, `V2PaginationImpl` will continue to iterate through
+all pages until the number of elements in the response no longer equals the expected count. 
+This will avoid infinitely looping over pages until an empty page is discovered.
+
+## 1.0.7 (Work in progress)